Time: 0 ms (100.00%) | Memory: 42.5 MB (24.68%) - LeetSync

+++ 3918-check-divisibility-by-digit-sum-and-product/check-divisibility-by-digit-sum-and-product.java
@@ -1,6 +1,6 @@
 class Solution {
     public boolean checkDivisibility(int n) {
-        long prod=1,sum=0;
+        int prod=1,sum=0;
         int num=n;
         while(n!=0){
             int rem=n%10;
